refactor(*): move streamer extension to their parent interface
