+++ core/src/main/java/io/github/thibaultbee/streampack/core/streamers/interfaces/ICameraStreamer.kt
@@ -15,8 +15,14 @@
  */
 package io.github.thibaultbee.streampack.core.streamers.interfaces
 
+import android.Manifest
 import android.view.Surface
+import android.view.SurfaceHolder
+import android.view.SurfaceView
+import android.view.TextureView
+import androidx.annotation.RequiresPermission
 import io.github.thibaultbee.streampack.core.internal.sources.video.camera.ICameraSource
+import io.github.thibaultbee.streampack.core.streamers.DefaultStreamer
 
 interface ICameraStreamer {
     /**
@@ -43,6 +49,29 @@ interface ICameraStreamer {
     fun stopPreview()
 }
 
+/**
+ * Sets a preview surface.
+ *
+ * @param surfaceView The [SurfaceView] used for camera preview
+ */
+fun ICameraStreamer.setPreview(surfaceView: SurfaceView) = setPreview(surfaceView.holder.surface)
+
+/**
+ * Sets a preview surface holder.
+ *
+ * @param surfaceHolder The [SurfaceHolder] used for camera preview
+ */
+fun ICameraStreamer.setPreview(surfaceHolder: SurfaceHolder) = setPreview(surfaceHolder.surface)
+
+/**
+ * Sets a preview surface.
+ *
+ * @param textureView The [TextureView] used for camera preview
+ */
+fun ICameraStreamer.setPreview(textureView: TextureView) =
+    setPreview(Surface(textureView.surfaceTexture))
+
+
 interface ICameraCoroutineStreamer : ICameraStreamer {
     /**
      * Starts audio and video capture.
@@ -50,9 +79,120 @@ interface ICameraCoroutineStreamer : ICameraStreamer {
     suspend fun startPreview()
 }
 
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param previewSurface The [Surface] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+suspend fun ICameraCoroutineStreamer.startPreview(previewSurface: Surface) {
+    setPreview(previewSurface)
+    startPreview()
+}
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param surfaceView The [SurfaceView] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+suspend fun ICameraCoroutineStreamer.startPreview(surfaceView: SurfaceView) =
+    startPreview(surfaceView.holder.surface)
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param surfaceHolder The [SurfaceHolder] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+suspend fun ICameraCoroutineStreamer.startPreview(surfaceHolder: SurfaceHolder) =
+    startPreview(surfaceHolder.surface)
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param textureView The [TextureView] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+suspend fun ICameraCoroutineStreamer.startPreview(textureView: TextureView) =
+    startPreview(Surface(textureView.surfaceTexture))
+
+
 interface ICameraCallbackStreamer : ICameraStreamer {
     /**
      * Starts audio and video capture.
      */
     fun startPreview()
 }
+
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param previewSurface The [Surface] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+fun ICameraCallbackStreamer.startPreview(previewSurface: Surface) {
+    setPreview(previewSurface)
+    startPreview()
+}
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param surfaceView The [SurfaceView] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+fun ICameraCallbackStreamer.startPreview(surfaceView: SurfaceView) =
+    startPreview(surfaceView.holder.surface)
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param surfaceHolder The [SurfaceHolder] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+fun ICameraCallbackStreamer.startPreview(surfaceHolder: SurfaceHolder) =
+    startPreview(surfaceHolder.surface)
+
+/**
+ * Starts audio and video capture.
+ * If you can prefer to call [DefaultStreamer.configure] before starting preview.
+ * It is a shortcut for [setPreview] and [startPreview].
+ *
+ * @param textureView The [TextureView] used for camera preview
+ *
+ * @see [stopPreview]
+ */
+@RequiresPermission(allOf = [Manifest.permission.CAMERA])
+fun ICameraCallbackStreamer.startPreview(textureView: TextureView) =
+    startPreview(Surface(textureView.surfaceTexture))
